# backend/src/swagger.json
{
  "openapi":  "3.0.0",
  "info": {
    "title": "Cyber Wallet",
    "description": "Documentation from cyber-wallet-ng api.",
    "version": "1.0.0"
  },
  "basePath": "/",
  "components": {
    "securitySchemes": {
      "Authorization": {
        "type": "apiKey",
        "in": "header",       
        "name": "Authorization"
      }
    }
  },
  "paths": {
    "/user": {
      "post": {
        "tags": ["Users"],
        "summary": "Create a User",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type":"string"
                  },
                  "password": {
                    "type":"string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "500": {
            "description": "Internal Server Error"
          },
          "400": {
            "description": "Bad Request"
          },
          "201": {
            "description": "User created"
          }
        }
      },
      "get": {
        "tags": ["Users"],
        "security":[{"Authorization": []}],
        "summary": "Get All Users",
        "responses": {
          "500": {
            "description": "Internal Server Error"
          },
          "401": {
            "description": "Unauthorized"
          },
          "200": {
            "description": "Users"
          }
        }
      }
    },
...
  }
}